+++ lib/src/cartesian/CartesianChart.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { type LayoutChangeEvent } from "react-native";
+import { View, type LayoutChangeEvent } from "react-native";
 import { Canvas, Group, type CanvasRef } from "@shopify/react-native-skia";
 import { useSharedValue } from "react-native-reanimated";
 import {
@@ -445,7 +445,16 @@ function CartesianChartContent<
             touchMap.value[touch.id] = i;
 
           v.isActive.value = true;
-          handleTouch(v, touch.x, touch.y);
+
+          // [3] and [7] are the X and Y translation components of the 4x4 transformation matrix, respectively.
+          const scrolledX = transformState?.matrix.value?.[3] || 0;
+          const scrolledY = transformState?.matrix.value?.[7] || 0;
+
+          handleTouch(
+            v,
+            touch.absoluteX - scrolledX,
+            touch.absoluteY - scrolledY,
+          );
         } else {
           gestureState.value.bootstrap.push([v, touch]);
         }
@@ -464,7 +473,15 @@ function CartesianChartContent<
           touchMap.value[touch.id] = i;
 
         v.isActive.value = true;
-        handleTouch(v, touch.x, touch.y);
+
+        const scrolledX = transformState?.matrix.value?.[3] || 0;
+        const scrolledY = transformState?.matrix.value?.[7] || 0;
+
+        handleTouch(
+          v,
+          touch.absoluteX - scrolledX,
+          touch.absoluteY - scrolledY,
+        );
       }
     })
     /**
@@ -489,7 +506,15 @@ function CartesianChartContent<
 
         if (!v || !touch) continue;
         if (!v.isActive.value) v.isActive.value = true;
-        handleTouch(v, touch.x, touch.y);
+
+        const scrolledX = transformState?.matrix.value?.[3] || 0;
+        const scrolledY = transformState?.matrix.value?.[7] || 0;
+
+        handleTouch(
+          v,
+          touch.absoluteX - scrolledX,
+          touch.absoluteY - scrolledY,
+        );
       }
     })
     /**
@@ -725,23 +750,23 @@ function CartesianChartContent<
   }
 
   return (
-    <GestureHandlerRootView
-      style={{ flex: 1, overflow: "hidden" }}
-      onLayout={onLayout}
-    >
-      {body}
-      <GestureHandler
-        config={gestureHandlerConfig}
-        gesture={composed}
-        transformState={transformState}
-        dimensions={{
-          x: Math.min(xScale.range()[0]!, 0),
-          y: Math.min(primaryYScale.range()[0]!, 0),
-          width: xScale.range()[1]! - Math.min(xScale.range()[0]!, 0),
-          height:
-            primaryYScale.range()[1]! - Math.min(primaryYScale.range()[0]!, 0),
-        }}
-      />
+    <GestureHandlerRootView>
+      <View style={{ flex: 1, overflow: "hidden" }} onLayout={onLayout}>
+        {body}
+        <GestureHandler
+          config={gestureHandlerConfig}
+          gesture={composed}
+          transformState={transformState}
+          dimensions={{
+            x: Math.min(xScale.range()[0]!, 0),
+            y: Math.min(primaryYScale.range()[0]!, 0),
+            width: xScale.range()[1]! - Math.min(xScale.range()[0]!, 0),
+            height:
+              primaryYScale.range()[1]! -
+              Math.min(primaryYScale.range()[0]!, 0),
+          }}
+        />
+      </View>
     </GestureHandlerRootView>
   );
 }