+++ example/CHANGELOG.md
@@ -1,5 +1,17 @@
 # example
 
+## 1.3.1
+
+### Patch Changes
+
+- all scrolling and scrubbing ([#633](https://github.com/FormidableLabs/victory-native-xl/pull/633))
+
+## 1.3.0
+
+### Minor Changes
+
+- Add `stepAfter` and `stepBefore` to the line curve options ([#618](https://github.com/FormidableLabs/victory-native-xl/pull/618))
+
 ## 1.2.3
 
 ### Patch Changes